+++ octgnFX/Octgn/Launcher/GameTableLauncher.cs
@@ -1,7 +1,9 @@
 ﻿namespace Octgn.Launcher
 {
     using System;
+    using System.Linq;
     using System.Net;
+    using System.Net.Sockets;
     using System.Reflection;
     using System.Threading;
     using System.Windows;
@@ -9,7 +11,9 @@
 
     using Octgn.Core.DataManagers;
     using Octgn.DataNew;
+    using Octgn.DataNew.Entities;
     using Octgn.Library.Exceptions;
+    using Octgn.Windows;
 
     using Skylabs.Lobby;
 
@@ -19,12 +23,30 @@
     {
         internal static ILog Log = LogManager.GetLogger(MethodBase.GetCurrentMethod().DeclaringType);
 
-        public void Launch()
+        internal int HostPort;
+        internal Game HostGame;
+        internal string HostUrl;
+
+        public void Launch(int? hostport, Guid? game)
         {
-            var gameId = Prefs.LastHostedGameType;
-            var game = GameManager.Get().GetById(gameId);
             Program.Dispatcher = Application.Current.Dispatcher;
-            StartLocalGame(game, Skylabs.Lobby.Randomness.RandomRoomName(), null);
+            HostGame = GameManager.Get().GetById(game.Value);
+            if (hostport == null || hostport <= 0)
+            {
+                this.HostPort = new Random().Next(5000, 6000);
+                while (!Networking.IsPortAvailable(this.HostPort)) this.HostPort++;
+            }
+            else
+            {
+                this.HostPort = hostport.Value;
+            }
+            // Host a game
+            this.Host();
+        }
+
+        private void Host()
+        {
+            StartLocalGame(HostGame, Skylabs.Lobby.Randomness.RandomRoomName(), null);
             Program.GameSettings.UseTwoSidedTable = Prefs.TwoSidedTable;
             if (Program.GameEngine != null)
                 Dispatcher.CurrentDispatcher.Invoke(new Action(Program.GameEngine.Begin));
@@ -40,14 +62,12 @@
 
         void StartLocalGame(DataNew.Entities.Game game, string name, string password)
         {
-            var hostport = new Random().Next(5000, 6000);
-            while (!Networking.IsPortAvailable(hostport)) hostport++;
-            var hs = new HostedGame(hostport, game.Id, game.Version, game.Name, name, null, new User(Prefs.Nickname + "@" + AppConfig.ChatServerPath), true);
+            var hs = new HostedGame(HostPort, game.Id, game.Version, game.Name, name, null, new User(Prefs.Nickname + "@" + AppConfig.ChatServerPath), true);
             if (!hs.StartProcess())
             {
                 throw new UserMessageException("Cannot start local game. You may be missing a file.");
             }
-            Program.LobbyClient.CurrentHostedGamePort = hostport;
+            Program.LobbyClient.CurrentHostedGamePort = HostPort;
             Program.GameSettings.UseTwoSidedTable = true;
             Program.GameEngine = new GameEngine(game, Prefs.Nickname, true);
             Program.IsHost = true;
@@ -58,7 +78,7 @@
             {
                 try
                 {
-                    Program.Client = new Octgn.Networking.Client(ip, hostport);
+                    Program.Client = new Octgn.Networking.Client(ip, HostPort);
                     Program.Client.Connect();
                     return;
                 }
@@ -70,6 +90,22 @@
                 Thread.Sleep(2000);
             }
             throw new UserMessageException("Cannot start local game. You may be missing a file.");
+        }
+
+        void JoinOfflineGame()
+        {
+            var addr = Dns.GetHostAddresses(HostUrl).First(x => x.AddressFamily == AddressFamily.InterNetwork);
+            Program.Client = new Octgn.Networking.Client(addr, HostPort);
+            Program.Client.Connect();
+
+            if (WindowManager.PreGameLobbyWindow == null)
+            {
+                Program.IsHost = false;
+                Program.GameEngine = new Octgn.GameEngine(HostGame, null, true);
+
+                WindowManager.PreGameLobbyWindow = new PreGameLobbyWindow();
+                WindowManager.PreGameLobbyWindow.Setup(true, null);
+            }
         }
     }
 }